pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seekr"
version = "0.1.0"
description = "Offline semantic file search engine — fully local, no cloud."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["search", "semantic", "embeddings", "faiss", "cli", "offline"]

# PyPI installs the CPU-only torch build by default (~200 MB).
# For CUDA, install torch separately before seekr:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
# then: pip install seekr
dependencies = [
    # CLI framework
    "typer[all]>=0.9.0",
    # Terminal output
    "rich>=13.0.0",
    # Vector store (CPU build; see [gpu] extra for GPU)
    "faiss-cpu>=1.7.4",
    # Text embeddings
    "sentence-transformers>=2.7.0",
    # Image embeddings (CLIP)
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "Pillow>=10.0.0",
    # File watching
    "watchdog>=4.0.0",
    # PDF support (optional but included for convenience)
    "pypdf>=4.0.0",
    # NumPy (required by sentence-transformers and faiss)
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.9.0",
    "ruff>=0.4.0",
    "black>=24.0.0",
]
# OCR support for scanned / image-only PDFs.
# Also requires the system tesseract binary:
#   Ubuntu/Debian: sudo apt install tesseract-ocr
#   macOS:         brew install tesseract
ocr = [
    "pdf2image>=1.17.0",
    "pytesseract>=0.3.10",
]
# GPU support: replaces faiss-cpu with faiss-gpu.
# Requires a CUDA-enabled torch installation (see note above).
gpu = [
    "faiss-gpu>=1.7.4",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/seekr"
"Bug Tracker" = "https://github.com/YOUR_USERNAME/seekr/issues"
Changelog = "https://github.com/YOUR_USERNAME/seekr/blob/main/CHANGELOG.md"

[project.scripts]
seekr = "seekr.cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["seekr*"]

# ---------------------------------------------------------------------------
# Black — code formatter
# ---------------------------------------------------------------------------

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

# ---------------------------------------------------------------------------
# Ruff — fast linter (replaces flake8, isort, pyupgrade, etc.)
# ---------------------------------------------------------------------------

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line-too-long (handled by formatter)
    "B008",   # do not perform function calls in default arguments (Typer pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["seekr"]

# ---------------------------------------------------------------------------
# Mypy — static type checker
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
# Allow untyped third-party decorators (Typer uses them)
disallow_untyped_decorators = false

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
